Move cuda_search into common tlcli binary
Most of the functionality of the `cuda_search` binary has been factored
out into the `telamon-cli` library.  It makes sense to move
`cuda_search` into a subcommand of the `tlcli` binary, as is the case
for the rest of the functionality, instead of using a separate binary.
