+++ telamon-cli/src/bin/tlcli.rs
@@ -23,7 +23,85 @@ use telamon::offline_analysis::tree::CandidateTree;
 use telamon::search_space::SearchSpace;
 use telamon_kernels::statistics::estimate_mean;
 
-use telamon_cli::{Bench, KernelBundle, KernelParam, Platform, ReplayPath};
+use telamon_cli::{Bench, CommonOpt, KernelBundle, KernelParam, Platform, ReplayPath};
+
+/// Run a full search for a given kernel
+#[derive(StructOpt)]
+struct Search {
+    #[structopt(flatten)]
+    common: CommonOpt,
+
+    #[structopt(short = "r", long = "repeat", default_value = "10")]
+    repeat: usize,
+
+    #[structopt(short = "k", long = "kernel")]
+    kernels: Vec<KernelParam>,
+
+    #[structopt(long = "platform", default_value = "cuda")]
+    platform: Platform,
+
+    /// Number of times to run the generated code to evaluate its performance.
+    #[structopt(long = "num-code-runs", default_value = "40")]
+    num_code_runs: usize,
+}
+
+impl Search {
+    fn run(&self, _args: &Opt) -> io::Result<()> {
+        let builder = self.platform.to_builder();
+        let mut config = self.common.config().unwrap().clone();
+        let output_base = std::path::Path::new(&config.output_dir).to_owned();
+
+        for idx in 0..self.repeat {
+            for kernel in &self.kernels {
+                config.output_dir = output_base
+                    .join(kernel.to_string())
+                    .join(idx.to_string())
+                    .to_str()
+                    .unwrap()
+                    .to_string();
+
+                let mut context = builder.build_context();
+                let (bundle, context) = context.kernel_bundle(kernel);
+
+                let best = explorer::find_best_ex(
+                    &config,
+                    context,
+                    bundle.candidates,
+                    Some({
+                        let check_fn = &bundle.check_fn;
+                        &move |_, context| check_fn(context)
+                    }),
+                )
+                .unwrap_or_else(|| panic!("no candidates found for kernel {}", kernel));
+
+                let best_fn = telamon::codegen::Function::build(&best.space);
+                let runtime = context.benchmark(&best_fn, self.num_code_runs);
+
+                let ref_runtime = Bench::default()
+                    .runs(self.num_code_runs)
+                    .benchmark_fn(&bundle.reference_fn);
+
+                let mut f =
+                    std::fs::File::create(config.output_path("benchmark.txt").unwrap())
+                        .unwrap();
+                writeln!(f, "runtimes: {:?}", runtime).unwrap();
+                let mean = estimate_mean(runtime, 0.95, "ns");
+                let ref_mean = estimate_mean(ref_runtime, 0.95, "ns");
+                writeln!(
+                    f,
+                    "{}: {}, reference: {}, speedup: {:.2}",
+                    kernel,
+                    mean,
+                    ref_mean,
+                    ref_mean.value / mean.value
+                )
+                .unwrap();
+            }
+        }
+
+        Ok(())
+    }
+}
 
 /// Compute the bound for a given candidate.
 #[derive(StructOpt)]
@@ -661,6 +739,9 @@ enum Command {
 
     #[structopt(name = "bound")]
     Bound(ComputeBound),
+
+    #[structopt(name = "search")]
+    Search(Search),
 }
 
 #[derive(StructOpt)]
@@ -681,6 +762,7 @@ fn main() {
         Command::Bounds(bounds) => bounds.run(&args),
         Command::Stats(stats) => stats.run(&args),
         Command::Bound(bound) => bound.run(&args),
+        Command::Search(search) => search.run(&args),
     };
 
     match result {
